Make it possible to restore presence on login
It uses the ConnectsAutomatically property and the AutomaticPresence.
In the kcm you can enable/disable this feature (ConnectsAutomatically
is set) and on every presence change the new presence is set as the
AutomaticPresence.
This way mission-control cares for setting the saved presence as soon as
possible.

If you want to manage this manually you can change the setting in
ktelepathyrc:

[KDED]
autoConnect=manual

BUG: 281929
BUG: 295842
REVIEW: 104363