Fixed the on-quit go offline? dialog box to always write to file.
Updated the on-quit go offline? dialog box to allow re-invocation when
it is set to never prompt again -- useful for when the go offline
checkbox is reset to manual. It previously didn't write to the file
when the user selected stay online -- because of this if the user
selects never ask me again and then go offline, then they get the
dialog back again through the new checkbox, and say stay online never
ask again, it doesn't use the correct choice.