Fixed the auto-offline checkbox to correctly not allow tristate.
The auto-offline checkbox will now not allow tristate settings when
the user has selected 'never ask me again' or checked/unchecked the
checkbox (that is when dont_check_for_plasmoid is true). It is only in
tristate before that happens when the checkbox will show up every time
the user exits.