%time Variable to get the current UTC time in the away message
To make this option visible I've changed the tool tip (I think the first thing a user would do if he wants no away message is to remove the message text [if there is one, which isn't by default], so this double information can be replaced).

Adding an extra widget for drag n drop like in now playing would be too much for this I think. It would blow up the UI if I add this under every lineedit (as more stuff is visible on the first sight for the user as more complicated it looks) and it would be confusing if I add it only once in a central place, so I decided to just inform the user through a tool tip (it's also not a important feature, so it's ok if it's a little bit hidden).

BUG: 300849
FIXED-IN: 0.7.0
REVIEW: 110334